+++ .gitignore
@@ -1,3 +1,4 @@
 .venv
+.vscode
 .ruff_cache
 poetry.lock
