+++ knovex/language/tnorm.py
