+++ knovex/reasoner/cqd.py
@@ -7,11 +7,11 @@ import torch
 from torch import nn
 from torch.nn import functional as F
 
-from ngdb.structure.kg_embedding.abstract_kge import (
+from knovex.structure.kg_embedding.abstract_kge import (
     KnowledgeGraphEmbedding as KGE,
 )
-from ngdb.reasoner.abstract_reasoner import Reasoner
-from ngdb.utils.data import VariadicMatrix, aggregation
+from knovex.reasoner.abstract_reasoner import Reasoner
+from knovex.utils.data import VariadicMatrix, aggregation
 
 logger = logging.getLogger(__name__)
