+++ knovex/structure/kg_embedding/conve.py
