+++ knovex/structure/kg_embedding/rescal.py
