+++ knovex/structure/kg_embedding/swtranse.py
