+++ knovex/structure/kg_embedding/transe.py
